[REFACTORING] Split IMAPServerAppendNonSynchronizedLitteralsTest
